fixed all package errors by renaming dir in test
