+++ src/test/java/flight_reservation/TestUtil.java
@@ -1,4 +1,4 @@
-package flight.reservation;
+package flight_reservation;
 
 import java.util.Calendar;
 import java.util.Date;
